Prevent crash when importing a watch-only wallet
When watch-only wallets are imported, the terminal will crash. The exact cause seems to have something to do with improperly establishing the LMDB environment for the watch-only wallet, with a write to the database failing during initial setup. Fix by catching a throw sent by LMDB when an error occurs.

Note that this commit is meant more to help with debugging and understanding the underlying issue. The wallet loads, including the current balance, but fails when attempting to do things like create new addresses.