Delete current wallet when importing a wallet
It's possible to have a full wallet on the system and not a watch-only wallet (or vice versa). The terminal currently stops an import if one of the other exists. Change the code to delete the currently existing wallet (if any exist) and re-import everything. This is done, in part, to ensure that changes to the wallets are picked up in both wallets (i.e., no sync issues).