Add more conditions for wallet-less Armory connections
After feedback, it was determined that the online condition needed to add a check for any wallets being present in the wallet manager. Armory should go online only if no wallets are present, including the settlement wallet (although the latter condition is commented out for now due to a bug that will be fixed elsewhere).