[BST-1835] Further tweak the RBF minimum fee
The original BST-1835 fix basically worked, with one caveat. The minimum fee was assumed to be the new fee/byte * the original TX virtual size. This isn't correct, and causes the RBF fee to potentially be higher than is strictly required. RBF policy essentially means the minimum fee is the original fee + the new TX virt size (1 sat / byte). Adjust accordingly so that the adjusted RBF fee is the minimum required.